change how RRD files are searched for
Add a switch/case statement that adjusts the glob() call based on the
plugin name.  This should make it more efficient so that
collectd_plugindata() only examines filenames that are of interest
rather then having to parse the name of every file every time.  For data
directories with a few hundred RRD files per host, this may save us a
lot of time.