centralize all file searching into rrd_file_search()
For the more complex plugins, we need the ability to change how files
are searched for and parsed.  Instead of doing that in two different
places, we are going to create a more generic function that looks at the
plugin name and adjusts the search paramenters.

This also allows us to search very specifically for a particular
plugin's files instead of searching for all RRD files in the host
directory.